+++ iconvdata/iso_6937.c
@@ -482,7 +482,7 @@ static const char from_ucs4[][2] =
     uint32_t ch = get32 (inptr);					      \
     const char *cp;							      \
 									      \
-    if (__builtin_expect (ch, 0)					      \
+    if ((size_t) __builtin_expect (ch, 0)				      \
 	>= sizeof (from_ucs4) / sizeof (from_ucs4[0]))			      \
       {									      \
 	int fail = 0;							      \
